requirements.txt
requests
feedparser
python-dotenv
PyQt6
